fix delete media list as well

+++ src/main/kotlin/ui/web/blink/domain/entities/DeleteMediaList.kt
@@ -8,5 +8,5 @@
 package ui.web.blink.domain.entities
 
 data class DeleteMediaList (
-    val mediaList: List<Int>
+    val mediaList: List<Long>
 )
